// apps/worker/src/archive/deprecated-components/vibegantt/types.ts
export interface DateRange {
  start: Date;
  end: Date;
}
// ...
export interface ViewportState {
  scrollX: number;
  scrollY: number;
  width: number;
  height: number;
  zoom: number;
  visibleDateRange: DateRange;
  visibleTaskIds: Set<string>;
}